Fix sonar checks

+++ src/lua/functions/creatures/npc/npc_type_functions.cpp
@@ -193,8 +193,8 @@ int NpcTypeFunctions::luaNpcTypeAddShopItem(lua_State* L) {
 	ShopInfo shopItem;
 
 	shopItem.itemClientId = static_cast<uint16_t>(getField<uint32_t>(L, table, "clientId"));
-	shopItem.buyPrice = static_cast<uint32_t>(getField<uint32_t>(L, table, "buy"));
-	shopItem.sellPrice = static_cast<uint32_t>(getField<uint32_t>(L, table, "sell"));
+	shopItem.buyPrice = getField<uint32_t>(L, table, "buy");
+	shopItem.sellPrice = getField<uint32_t>(L, table, "sell");
 	shopItem.subType = static_cast<int32_t>(getField<uint32_t>(L, table, "count"));
 	shopItem.storageKey = static_cast<int32_t>(getField<uint32_t>(L, table, "storageKey"));
 	shopItem.storageValue = static_cast<int32_t>(getField<uint32_t>(L, table, "storageValue"));
